Reajuste de las practicas del mòdulo 4, donde dividimos el còdigo en los archivos tipo clase correspondientes
