+++ practice_4_2/src/main/kotlin/EBook.kt
@@ -1,0 +1,17 @@
+package org.example
+
+
+class EBook(title: String, author: String, val format: String = "text") : Book(title, author) {
+    // Variable privada para almacenar el conteo de palabras
+    private var wordCount = 0
+
+    // Sobrescribir el método readPage para incrementar el conteo de palabras
+    override fun readPage() {
+        wordCount += 250
+    }
+
+    // Método para obtener el conteo de palabras (opcional para verificación)
+    fun getWordCount(): Int {
+        return wordCount
+    }
+}
